feat: extract components
