+++ src/components/App.js
@@ -1,23 +1,19 @@
 import React, { useEffect, useState } from "react";
 import "antd/dist/antd.css";
 import "./App.css";
-import { Layout, Menu } from "antd";
-import {
-  ExperimentOutlined,
-  ForkOutlined,
-  UserOutlined,
-  TeamOutlined,
-} from "@ant-design/icons";
-import { Redirect, Route, Link, Switch } from "react-router-dom";
-import BeaconOverview from "./components/BeaconOverview";
-import QueryVariants from "./components/QueryVariants";
-import QueryBiosamples from "./components/QueryBiosamples";
-import QueryIndividuals from "./components/QueryIndividuals";
-import QueryResults from "./components/QueryResults";
-import { apiGet } from "./api";
+import { Layout } from "antd";
+import { Route, Link, Switch } from "react-router-dom";
+import Sidebar from "./Sidebar";
+import Header from "./Header";
+import Footer from "./Footer";
+import BeaconOverview from "./BeaconOverview";
+import QueryVariants from "./QueryVariants";
+import QueryBiosamples from "./QueryBiosamples";
+import QueryIndividuals from "./QueryIndividuals";
+import QueryResults from "./QueryResults";
+import { apiGet } from "../api";
 
 const App = () => {
-  const { Header, Content, Footer, Sider } = Layout;
   const [beaconInfo, setBeaconInfo] = useState(null);
   const [queryData, setQueryData] = useState({});
   const [apiRoute, setApiRoute] = useState('')
@@ -50,41 +46,10 @@ const App = () => {
 
   return (
     <Layout>
-      <Sider
-        breakpoint="lg"
-        collapsedWidth="0"
-        onBreakpoint={(broken) => {
-          // console.log(broken);
-        }}
-        onCollapse={(collapsed, type) => {
-          // console.log(collapsed, type);
-        }}
-      >
-        <div className="logo" />
-        <Menu theme="dark" mode="inline" defaultSelectedKeys={["4"]}>
-          <Menu.Item key="1" icon={<UserOutlined />}>
-            <Link to="/overview">Overview</Link>
-          </Menu.Item>
-          <Menu.Item key="2" icon={<ForkOutlined />}>
-            <Link to="/query-variants">Query Variants</Link>
-          </Menu.Item>
-          <Menu.Item key="3" icon={<ExperimentOutlined />}>
-            <Link to="/query-biosamples">Query Biosamples</Link>
-          </Menu.Item>
-          <Menu.Item key="4" icon={<TeamOutlined />}>
-            <Link to="/query-individuals">Query Individuals</Link>
-          </Menu.Item>
-        </Menu>
-      </Sider>
+      <Sidebar />
       <Layout>
-        <Header
-          className="site-layout-sub-header-background"
-          style={{ padding: 0 }}
-        >
-          {" "}
-          Header{" "}
-        </Header>
-        <Content style={{ margin: "24px 16px 0" }}>
+        <Header />
+        <Layout.Content style={{ margin: "24px 16px 0" }}>
           <div
             className="site-layout-background"
             style={{ padding: 24, minHeight: 360 }}
@@ -120,8 +85,8 @@ const App = () => {
               <Route path="">404: Nothing here</Route>
             </Switch>
           </div>
-        </Content>
-        <Footer style={{ textAlign: "center" }}>bottom text</Footer>
+        </Layout.Content>
+        <Footer />
       </Layout>
     </Layout>
   );
