+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { Layout } from "antd";
+
+export default function Footer() {
+  return (
+    <Layout.Footer style={{ textAlign: "center" }}>
+      footer
+    </Layout.Footer>
+  )
+}
+
