+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Layout } from "antd";
+
+export default function Header() {
+  return (
+    <Layout.Header
+      className="site-layout-sub-header-background"
+      style={{ padding: 0 }}
+    >
+      {" "}
+      Header{" "}
+    </Layout.Header>
+  )
+}
